feat: bump version to 0.6

+++ docs/tutorial.py
@@ -78,7 +78,7 @@ plt.imshow(zarr_arr, cmap="gray")
 array_specs: list[AnyArraySpec] = [
     ArraySpec(
         shape=(100, 100),
-        data_type=np.uint16,
+        data_type=np.uint8,
         chunk_grid=NamedConfig(
             name="regular",
             configuration={"chunk_shape": [32, 32]},
@@ -93,7 +93,7 @@ array_specs: list[AnyArraySpec] = [
     ),
     ArraySpec(
         shape=(100, 100),
-        data_type=np.uint16,
+        data_type=np.uint8,
         chunk_grid=NamedConfig(
             name="regular",
             configuration={"chunk_shape": [32, 32]},
@@ -131,14 +131,14 @@ pprint(ome_zarr_image)
 #
 # For numpy arrays:
 
-arr0 = np.zeros(shape=(100, 100), dtype=np.uint16)
-arr1 = np.zeros(shape=(50, 50), dtype=np.uint16)
+arr0 = np.zeros(shape=(100, 100), dtype=np.uint8)
+arr1 = np.zeros(shape=(50, 50), dtype=np.uint8)
 array_specs = [ArraySpec.from_array(arr0), ArraySpec.from_array(arr1)]
 
 # or for Zarr arrays:
 
-arr_zarr0 = zarr.zeros(shape=(100, 100), dtype=np.uint16, zarr_format=2)
-arr_zarr1 = zarr.zeros(shape=(50, 50), dtype=np.uint16, zarr_format=2)
+arr_zarr0 = zarr.zeros(shape=(100, 100), dtype=np.uint8, zarr_format=2)
+arr_zarr1 = zarr.zeros(shape=(50, 50), dtype=np.uint8, zarr_format=2)
 array_specs = [ArraySpec.from_array(arr_zarr0), ArraySpec.from_array(arr_zarr1)]
 
 # ## Saving datasets
